sentiment-api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
google-generativeai==0.8.3
requests==2.31.0
python-dotenv==1.0.0

